+++ udata/core/dataset/metrics.py
@@ -9,12 +9,11 @@ from udata.models import Dataset, Reuse
 from udata.core.issues.metrics import IssuesMetric
 from udata.core.discussions.metrics import DiscussionsMetric
 from udata.core.followers.metrics import FollowersMetric
-from udata.core.badges.metrics import BadgesMetric
 
 
 __all__ = (
     'DatasetReuses', 'DatasetIssuesMetric', 'DatasetFollowers',
-    'DatasetDiscussionsMetric', 'DatasetBadgesMetric'
+    'DatasetDiscussionsMetric'
 )
 
 
@@ -44,8 +43,4 @@ class DatasetDiscussionsMetric(DiscussionsMetric):
 
 
 class DatasetFollowers(FollowersMetric):
-    model = Dataset
-
-
-class DatasetBadgesMetric(BadgesMetric):
     model = Dataset