+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   [#882](https://github.com/opendatateam/udata/issues/882)
 - Proper producer logo display on dataset pages
 - Fix CKAN harvester empty notes and `metadata` file type handling
+- Remove (temporary) badges metrics
+  [#885](https://github.com/opendatateam/udata/issues/885)
 
 ## 1.0.9 (2017-04-23)